+++ lib/collection.ts
@@ -13,24 +13,25 @@
 // limitations under the License.
 
 import { WebElement } from 'selenium-webdriver';
-import { CollectionCondition, ElementCondition } from './conditions';
+import { ElementCondition } from './conditions';
 import { Configuration } from './configuration';
 import { Element } from './element';
-import { ByIndexWebElementLocator } from './locators/byIndexWebElementLocator';
-import { CashedWebElementLocator } from './locators/cashedWebElementLocator';
-import { FilteredByConditionWebElementsLocator } from './locators/filteredByConditionWebElementsLocator';
-import { Locator } from './locators/locator';
-import { Condition, Query, Wait } from './wait';
 import { Assertable, Entity, Matchable } from './entity';
 import { ElementByConditionWebElementLocator } from './locators/byConditionWebElementLocator';
-import { have } from './support/conditions/have';
+import { ByIndexWebElementLocator } from './locators/byIndexWebElementLocator';
+import { CashedWebElementLocator } from './locators/cashedWebElementLocator';
+import { CollectedByLocator } from './locators/collectedByLocator';
+import { FilteredByConditionWebElementsLocator } from './locators/filteredByConditionWebElementsLocator';
+import { Locator } from './locators/locator';
 import { SlicedWebElementsLocator } from './locators/slicedWebElementsLocator';
+import { Condition } from './wait';
+
 
 export class Collection extends Entity implements Assertable, Matchable {
 
-    constructor(private readonly locator: Locator<Promise<WebElement[]>>,
-                protected readonly configuration: Configuration) {
-                // readonly configuration: Configuration) {
+    private readonly locator: Locator<Promise<WebElement[]>>;
+
+    constructor(locator: Locator<Promise<WebElement[]>>, configuration: Configuration) {
         super(configuration);
         this.locator = locator;
     }
@@ -76,6 +77,10 @@ export class Collection extends Entity implements Assertable, Matchable {
             new ElementByConditionWebElementLocator(Condition.all(...conditions), this),
             this.configuration
         );
+    }
+
+    collected(searchFunction: (element: Element) => Element | Collection): Collection {
+        return new Collection(new CollectedByLocator(searchFunction, this), this.configuration);
     }
 
     async getWebElements(): Promise<WebElement[]> {