+++ lib/utils/extensions.ts
@@ -14,8 +14,11 @@
 
 import * as fs from 'fs-extra';
 import { By, WebDriver } from 'selenium-webdriver';
-import { Browser } from '../browser';
+import { Locator } from '../locators/locator';
 import { by } from '../support/selectors/by';
+import { Browser } from '..';
+import { BrowserWebElementByLocator } from '../locators/BrowserWebElementByLocator';
+import { ElementWebElementByLocator } from '../locators/ElementWebElementByLocator';
 
 
 export namespace Extensions {
@@ -42,8 +45,12 @@ export namespace Extensions {
     }
 
     export function toBy(cssOrXpathOrBy: string | By): By {  // todo: probably we need to enhance xpath identification
+        const isXpath = (str: string) => str.startsWith('/')
+        || str.startsWith('./')
+        || str.startsWith('..')
+        || str.startsWith('(');
         return (typeof cssOrXpathOrBy === 'string')
-            ? cssOrXpathOrBy.includes('/') ? by.xpath(cssOrXpathOrBy) : by.css(cssOrXpathOrBy)
+            ? isXpath(cssOrXpathOrBy) ? by.xpath(cssOrXpathOrBy) : by.css(cssOrXpathOrBy)
             : cssOrXpathOrBy;
     }